\title{Graphical User Interface for the limma microarray package by Gordon Smyth}
\name{limmaGUI}
\alias{limmaGUI}
\alias{GetComponentsToExportInHTMLreport}
\alias{HTMLplotUsingFunction}
\alias{ExportHTMLreport}
\alias{TryReadImgProcFile}
\alias{Require}
\alias{TclRequire}
\alias{getPackageVersion}
\alias{initGlobals}
\alias{deleteItemFromList}
\alias{SetLayoutParameters}
\alias{OpenGALFile}
\alias{OpenTargetsFile}
\alias{OpenSpotTypesFile}
\alias{GetImageProcessingFileType}
\alias{ReadImageProcessingFiles}
\alias{tclArrayVar}
\alias{GetDesignOrContrasts}
\alias{SimplifyContrastsExpression}
\alias{GetRNATypesFrom.ContrastsFromDropDowns.String}
\alias{GetParameterNames}
\alias{GetReducedDuplicateSpacing}
\alias{GetContrastsParameterizationNames}
\alias{GetSpotTypesIncludedNames}
\alias{HowManyDups}
\alias{ViewDesignOrContrastsMatrixInTable}
\alias{ViewDesignOrContrastsMatrixAsPairs}
\alias{ViewExistingContrastsParameterization}
\alias{ViewExistingParameterization}
\alias{CreateNewParameterization}
\alias{ViewRNATargets}
\alias{nstrstr}
\alias{strstr}
\alias{ViewSpotTypes}
\alias{ComputeContrasts}
\alias{ComputeLinearModel}
\alias{ChooseParameterization}
\alias{GetCoef}
\alias{showTopTable}
\alias{GetSpotTypesForLinearModel}
\alias{GetSlideNum}
\alias{GetDEcutoff}
\alias{UpdateSpotTypesStatus}
\alias{ChooseEbayesStatistic}
\alias{ChooseSpotType}
\alias{GetWtAreaParams}
\alias{GetWtFlagParams}
\alias{evalRcode}
\alias{OpenGALandTargetsandSpotTypesfiles}
\alias{GetlimmaDataSetName}
\alias{GetParameterizationName}
\alias{GetContrastsParameterizationName}
\alias{NewLimmaFile}
\alias{SetWD}
\alias{onExit}
\alias{DeleteParameterization}
\alias{ChooseContrastsParameterization}
\alias{DeleteContrastsParameterization}
\alias{OpenLimmaFile}
\alias{SaveLimmaFile}
\alias{SaveAsLimmaFile}
\alias{AboutLimmaGUI}
\alias{addMenuItem}
\alias{GetMenuName}
\alias{AboutCustomMenu}
\alias{DeleteCustomMenuItem}
\alias{GetCustomMenuItemNum}
\alias{GetNormalizationMethod}
\alias{NormalizeNow}
\alias{AboutNormalization}
\alias{GetGeneLabelsOptions}
\alias{MAPlotAvg}
\alias{MMPlot}
\alias{GetParametersAndOrContrasts}
\alias{HeatDiagramDialog}
\alias{HeatDiagramPlot}
\alias{ImageArrayPlotDialog}
\alias{ImageArrayPlot}
\alias{LogOddsPlot}
\alias{DupCorBoxPlot}
\alias{QQTplot}
\alias{MBoxPlot}
\alias{PrintTipGroupMAPlot}
\alias{MAPlot}
\alias{ChoosePlotSymbolByClicking}
\alias{SelectPlotSymbols}
\alias{plotMAColorCoded}
\alias{GetNEWxlim}
\alias{ebayesBoxPlots}
\alias{GetPlotLabels}
\alias{GetPlotTitle}
\alias{GetPlotSize}
\alias{SaveGraphAsJpeg}
\alias{SaveGraphAsPNG}
\alias{SaveGraphAsPostscript}
\alias{SaveGraphAsPDF}
\alias{Resize}
\alias{CopyGraph}
\alias{SetupPlotKeyBindings}
\alias{SetupPlotMenus}
\alias{GetJpegOrPngParams}
\alias{GetJpegOrPngX11Params}
\alias{VennDiagramPlot}
\alias{UpDownOrBoth}
\alias{chooseDir}
\alias{ComputeLinearModelFit}
\alias{CreateTreeAndList}
\alias{ExportAvalues}
\alias{ExportMvalues}
\alias{GetGenePixFlagWeightings}
\alias{imageplotlimmaGUI}
\alias{InitNewParameterization}
\alias{limmaGUIhelp}
\alias{limmaHelp}
\alias{plotMAlimmaGUI}
\alias{readGALlimmaGUI}
\alias{Try}
\alias{vennDiagramlimmaGUI}
\alias{WithinBetweenOrBoth}
\alias{wtflags2}

\description{Graphical User Interface for the limma microarray package by Gordon Smyth}}

\usage{
limmaGUI()
GetComponentsToExportInHTMLreport()
HTMLplotUsingFunction()
ExportHTMLreport()
TryReadImgProcFile()
Require()
TclRequire()
getPackageVersion()
initGlobals()
deleteItemFromList()
SetLayoutParameters()
OpenGALFile()
OpenTargetsFile()
OpenSpotTypesFile()
GetImageProcessingFileType()
ReadImageProcessingFiles()
tclArrayVar()
GetDesignOrContrasts()
SimplifyContrastsExpression()
GetRNATypesFrom.ContrastsFromDropDowns.String()
GetParameterNames()
GetReducedDuplicateSpacing()
GetContrastsParameterizationNames()
GetSpotTypesIncludedNames()
HowManyDups()
ViewDesignOrContrastsMatrixInTable()
ViewDesignOrContrastsMatrixAsPairs()
ViewExistingContrastsParameterization()
ViewExistingParameterization()
CreateNewParameterization()
ViewRNATargets()
nstrstr()
strstr()
ViewSpotTypes()
ComputeContrasts()
ComputeLinearModel()
ChooseParameterization()
GetCoef()
showTopTable()
GetSpotTypesForLinearModel()
GetSlideNum()
GetDEcutoff()
UpdateSpotTypesStatus()
ChooseEbayesStatistic()
ChooseSpotType()
GetWtAreaParams()
GetWtFlagParams()
evalRcode()
OpenGALandTargetsandSpotTypesfiles()
GetlimmaDataSetName()
GetParameterizationName()
GetContrastsParameterizationName()
NewLimmaFile()
SetWD()
onExit()
DeleteParameterization()
ChooseContrastsParameterization()
DeleteContrastsParameterization()
OpenLimmaFile()
SaveLimmaFile()
SaveAsLimmaFile()
AboutLimmaGUI()
addMenuItem()
GetMenuName()
AboutCustomMenu()
DeleteCustomMenuItem()
GetCustomMenuItemNum()
GetNormalizationMethod()
NormalizeNow()
AboutNormalization()
GetGeneLabelsOptions()
MAPlotAvg()
MMPlot()
GetParametersAndOrContrasts()
HeatDiagramDialog()
HeatDiagramPlot()
ImageArrayPlotDialog()
ImageArrayPlot()
LogOddsPlot()
DupCorBoxPlot()
QQTplot()
MBoxPlot()
PrintTipGroupMAPlot()
MAPlot()
ChoosePlotSymbolByClicking()
SelectPlotSymbols()
plotMAColorCoded()
GetNEWxlim()
ebayesBoxPlots()
GetPlotLabels()
GetPlotTitle()
GetPlotSize()
SaveGraphAsJpeg()
SaveGraphAsPNG()
SaveGraphAsPostscript()
SaveGraphAsPDF()
Resize()
CopyGraph()
SetupPlotKeyBindings()
SetupPlotMenus()
GetJpegOrPngParams()
GetJpegOrPngX11Params()
VennDiagramPlot()
UpDownOrBoth()

limmaGUI(BigfontsForlimmaGUIpresentation=FALSE)
}

\arguments{
	\item{BigfontsForlimmaGUIpresentation}{If set to TRUE, larger fonts are used.  However, some font 
	sizes are not controlled by limmaGUI and so must be adjusted in the operating system, e.g.
	in the Control Panel in Windows under Display, Appearance.}
}

\details{
This function launches a Graphical User Interface for the limma package by Gordon Smyth.
The GUI uses Tk widgets (via the R TclTk interface by Peter Dalgaard) in order to provide
a simple interface to the limma functions for linear modelling of microarrays and 
identification of differentially expressed genes.
}

\author{James Wettenhall}

\keyword{IO}
